requirements.txt
pillow==10.1.0
django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
django-filter==23.5
pre-commit==3.6.0
